# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "name": "Holiday Extra",
    "description": "API Tech test"
  },
  "definitions": {
    "Success": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "userId": {
      "properties": {
        "uuid": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "givenName": {
          "type": "string"
        },
        "familyName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "FullUser": {
      "type": "object",
      "properties": {
        "givenName": {
          "type": "string"
        },
        "familyName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "created": {
          "type": "string"
        }
      }
    },
    "Users": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/FullUser"
      }
    }
  },
  "tags": [
    {
      "name": "users",
      "description": "api for users in the sysytems"
    }
  ],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/user": {
      "post": {
        "tags": ["User"],
        "summary": "Create a new user in system",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "The User we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "requestBody": {
          "description": "User Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FullUser"
            }
          },
          "400": {
            "description": "User Already Exists"
          }
        }
      }
    },
    "/users": {
      "parameters": [],
      "get": {
        "tags": ["Users"],
        "summary": "Get all users in system",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "404": {
            "description": "Users not found"
          }
        }
      }
    },
    "/user/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID of the user that we want to match",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["User"],
        "summary": "Get user with given ID",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "User with id",
            "schema": {
              "$ref": "#/definitions/userId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FullUser"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "summary": "Update user with given ID",
        "tags": ["User"],
        "requestBody": {
          "description": "User Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "User with new values of properties",
            "schema": {
              "$ref": "#/definitions/userId"
            }
          },
          {
            "in": "body",
            "name": "body",
            "description": "The User we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FullUser"
            }
          },
          "400": {
            "description": "User not found."
          }
        }
      },
      "delete": {
        "summary": "Delete user with given ID",
        "tags": ["User"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "Delete user with id",
            "schema": {
              "$ref": "#/definitions/userId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          },
          "404": {
            "description": "Failed. User not found."
          }
        }
      }
    }
  }
}